v2.0/web/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
requests>=2.31.0
